move common fo compare scripts tools to a separate file. Add shift-function for plotting. Add FillSampledic function
